requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
certifi==2022.5.18.1
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cssselect2==0.6.0
cycler==0.11.0
Django==4.0.4
django-ckeditor==6.4.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-htmx==1.13.0
django-js-asset==2.0.0
django-smartfields==1.1.3
djangorestframework==3.13.1
fonttools==4.33.3
future==0.18.2
html5lib==1.1
idna==3.3
joblib==1.1.0
kiwisolver==1.4.2
lxml==4.9.1
matplotlib==3.5.2
numpy==1.22.3
oscrypto==1.3.0
packaging==21.3
pandas==1.4.2
pep8==1.7.1
Pillow==9.3.0
pycodestyle==2.5.0
pycparser==2.21
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyJWT==2.6.0
pyparsing==3.0.9
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.0
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
svglib==1.4.1
threadpoolctl==3.1.0
tinycss2==1.1.1
tqdm==4.64.1
twilio==7.15.1
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
xgboost==1.5.0
xhtml2pdf==0.2.8
